PMM-2214 Fixed default value for ExampleQuery conf
Since Go default for a boolean var is 'false', there was no way to tell
if the key ExampleQueries was missing in the config file or if it was
set to false.
The agent was not validating that key and it was setting it to true
instead. I changed that variable as a pointer to a boolean.
Now, if the key is missing in the config file, the variable will be nil
and in that case we can set it to the default=true, otherwise, it will
be left as set into the config file.